third_party/blink/web_tests/external/wpt/css/css-values/calc-size/animation/interpolate-size-max-height-interpolation-expected.txt

This is a testharness.js-based test.
Found 66 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <max-height> from [stretch] to [fit-content] at (-0.3) should be [stretch]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <max-height> from [stretch] to [fit-content] at (0) should be [stretch]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <max-height> from [stretch] to [fit-content] at (0.3) should be [stretch]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <max-height> from [stretch] to [fit-content] at (0.5) should be [fit-content]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <max-height> from [stretch] to [fit-content] at (0.6) should be [fit-content]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <max-height> from [stretch] to [fit-content] at (1) should be [fit-content]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <max-height> from [stretch] to [fit-content] at (1.5) should be [fit-content]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-height> from [stretch] to [fit-content] at (-0.3) should be [stretch]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-height> from [stretch] to [fit-content] at (0) should be [stretch]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-height> from [stretch] to [fit-content] at (0.3) should be [stretch]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-height> from [stretch] to [fit-content] at (0.5) should be [fit-content]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-height> from [stretch] to [fit-content] at (0.6) should be [fit-content]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-height> from [stretch] to [fit-content] at (1) should be [fit-content]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-height> from [stretch] to [fit-content] at (1.5) should be [fit-content]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Transitions: property <max-height> from [stretch] to [fit-content] at (-0.3) should be [fit-content]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Transitions: property <max-height> from [stretch] to [fit-content] at (0) should be [fit-content]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Transitions: property <max-height> from [stretch] to [fit-content] at (0.3) should be [fit-content]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Transitions: property <max-height> from [stretch] to [fit-content] at (0.5) should be [fit-content]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Transitions: property <max-height> from [stretch] to [fit-content] at (0.6) should be [fit-content]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Transitions: property <max-height> from [stretch] to [fit-content] at (1) should be [fit-content]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Transitions: property <max-height> from [stretch] to [fit-content] at (1.5) should be [fit-content]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Transitions with transition: all: property <max-height> from [stretch] to [fit-content] at (-0.3) should be [fit-content]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Transitions with transition: all: property <max-height> from [stretch] to [fit-content] at (0) should be [fit-content]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Transitions with transition: all: property <max-height> from [stretch] to [fit-content] at (0.3) should be [fit-content]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Transitions with transition: all: property <max-height> from [stretch] to [fit-content] at (0.5) should be [fit-content]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Transitions with transition: all: property <max-height> from [stretch] to [fit-content] at (0.6) should be [fit-content]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Transitions with transition: all: property <max-height> from [stretch] to [fit-content] at (1) should be [fit-content]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Transitions with transition: all: property <max-height> from [stretch] to [fit-content] at (1.5) should be [fit-content]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Animations: property <max-height> from [stretch] to [fit-content] at (-0.3) should be [stretch]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Animations: property <max-height> from [stretch] to [fit-content] at (0) should be [stretch]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Animations: property <max-height> from [stretch] to [fit-content] at (0.3) should be [stretch]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Animations: property <max-height> from [stretch] to [fit-content] at (0.5) should be [fit-content]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Animations: property <max-height> from [stretch] to [fit-content] at (0.6) should be [fit-content]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Animations: property <max-height> from [stretch] to [fit-content] at (1) should be [fit-content]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Animations: property <max-height> from [stretch] to [fit-content] at (1.5) should be [fit-content]
  assert_true: 'from' value should be supported expected true got false
[FAIL] Web Animations: property <max-height> from [stretch] to [fit-content] at (-0.3) should be [stretch]
  assert_true: 'from' value should be supported expected true got false
[FAIL] Web Animations: property <max-height> from [stretch] to [fit-content] at (0) should be [stretch]
  assert_true: 'from' value should be supported expected true got false
[FAIL] Web Animations: property <max-height> from [stretch] to [fit-content] at (0.3) should be [stretch]
  assert_true: 'from' value should be supported expected true got false
[FAIL] Web Animations: property <max-height> from [stretch] to [fit-content] at (0.5) should be [fit-content]
  assert_true: 'from' value should be supported expected true got false
[FAIL] Web Animations: property <max-height> from [stretch] to [fit-content] at (0.6) should be [fit-content]
  assert_true: 'from' value should be supported expected true got false
[FAIL] Web Animations: property <max-height> from [stretch] to [fit-content] at (1) should be [fit-content]
  assert_true: 'from' value should be supported expected true got false
[FAIL] Web Animations: property <max-height> from [stretch] to [fit-content] at (1.5) should be [fit-content]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Animations: property <max-height> from [min-content] to neutral at (-0.3) should be [calc-size(min-content, -30px + size * 1.3)]
  assert_equals: expected "calc - size ( min - content , - 30px + size * 1.3 ) " but got "calc - size ( min - content , - 30px + 0px + size * 1.3 ) "
[FAIL] CSS Animations: property <max-height> from [min-content] to neutral at (0) should be [calc-size(min-content, 0px + size * 1)]
  assert_equals: expected "calc - size ( min - content , 0px + size * 1 ) " but got "calc - size ( min - content , 0px + 0px + size * 1 ) "
[FAIL] CSS Animations: property <max-height> from [min-content] to neutral at (0.3) should be [calc-size(min-content, 30px + size * 0.7)]
  assert_equals: expected "calc - size ( min - content , 30px + size * 0.7 ) " but got "calc - size ( min - content , 30px + 0px + size * 0.7 ) "
[FAIL] CSS Animations: property <max-height> from [min-content] to neutral at (0.6) should be [calc-size(min-content, 60px + size * 0.4)]
  assert_equals: expected "calc - size ( min - content , 60px + size * 0.4 ) " but got "calc - size ( min - content , 60px + 0px + size * 0.4 ) "
[FAIL] CSS Animations: property <max-height> from [min-content] to neutral at (1) should be [calc-size(min-content, 100px + size * 0)]
  assert_equals: expected "calc - size ( min - content , 100px + size * 0 ) " but got "calc - size ( min - content , 100px + 0px + size * 0 ) "
[FAIL] CSS Animations: property <max-height> from [min-content] to neutral at (1.5) should be [calc-size(min-content, 150px + size * -0.5)]
  assert_equals: expected "calc - size ( min - content , 150px + size * - 0.5 ) " but got "calc - size ( min - content , 150px + 0px + size * - 0.5 ) "
[FAIL] Web Animations: property <max-height> from [min-content] to neutral at (-0.3) should be [calc-size(min-content, -30px + size * 1.3)]
  assert_equals: expected "calc - size ( min - content , - 30px + size * 1.3 ) " but got "calc - size ( min - content , - 30px + 0px + size * 1.3 ) "
[FAIL] Web Animations: property <max-height> from [min-content] to neutral at (0) should be [calc-size(min-content, 0px + size * 1)]
  assert_equals: expected "calc - size ( min - content , 0px + size * 1 ) " but got "calc - size ( min - content , 0px + 0px + size * 1 ) "
[FAIL] Web Animations: property <max-height> from [min-content] to neutral at (0.3) should be [calc-size(min-content, 30px + size * 0.7)]
  assert_equals: expected "calc - size ( min - content , 30px + size * 0.7 ) " but got "calc - size ( min - content , 30px + 0px + size * 0.7 ) "
[FAIL] Web Animations: property <max-height> from [min-content] to neutral at (0.6) should be [calc-size(min-content, 60px + size * 0.4)]
  assert_equals: expected "calc - size ( min - content , 60px + size * 0.4 ) " but got "calc - size ( min - content , 60px + 0px + size * 0.4 ) "
[FAIL] Web Animations: property <max-height> from [min-content] to neutral at (1) should be [calc-size(min-content, 100px + size * 0)]
  assert_equals: expected "calc - size ( min - content , 100px + size * 0 ) " but got "calc - size ( min - content , 100px + 0px + size * 0 ) "
[FAIL] Web Animations: property <max-height> from [min-content] to neutral at (1.5) should be [calc-size(min-content, 150px + size * -0.5)]
  assert_equals: expected "calc - size ( min - content , 150px + size * - 0.5 ) " but got "calc - size ( min - content , 150px + 0px + size * - 0.5 ) "
[FAIL] CSS Animations: property <max-height> from neutral to [max-content] at (-0.3) should be [calc-size(max-content, 130px + size * -0.3)]
  assert_equals: expected "calc - size ( max - content , 130px + size * - 0.3 ) " but got "calc - size ( max - content , 130px + 0px + size * - 0.3 ) "
[FAIL] CSS Animations: property <max-height> from neutral to [max-content] at (0) should be [calc-size(max-content, 100px + size * 0)]
  assert_equals: expected "calc - size ( max - content , 100px + size * 0 ) " but got "calc - size ( max - content , 100px + 0px + size * 0 ) "
[FAIL] CSS Animations: property <max-height> from neutral to [max-content] at (0.3) should be [calc-size(max-content, 70px + size * 0.3)]
  assert_equals: expected "calc - size ( max - content , 70px + size * 0.3 ) " but got "calc - size ( max - content , 70px + 0px + size * 0.3 ) "
[FAIL] CSS Animations: property <max-height> from neutral to [max-content] at (0.6) should be [calc-size(max-content, 40px + size * 0.6)]
  assert_equals: expected "calc - size ( max - content , 40px + size * 0.6 ) " but got "calc - size ( max - content , 40px + 0px + size * 0.6 ) "
[FAIL] CSS Animations: property <max-height> from neutral to [max-content] at (1) should be [calc-size(max-content, 0px + size * 1)]
  assert_equals: expected "calc - size ( max - content , 0px + size * 1 ) " but got "calc - size ( max - content , 0px + 0px + size * 1 ) "
[FAIL] CSS Animations: property <max-height> from neutral to [max-content] at (1.5) should be [calc-size(max-content, -50px + size * 1.5)]
  assert_equals: expected "calc - size ( max - content , - 50px + size * 1.5 ) " but got "calc - size ( max - content , - 50px + 0px + size * 1.5 ) "
[FAIL] Web Animations: property <max-height> from neutral to [max-content] at (-0.3) should be [calc-size(max-content, 130px + size * -0.3)]
  assert_equals: expected "calc - size ( max - content , 130px + size * - 0.3 ) " but got "calc - size ( max - content , 130px + 0px + size * - 0.3 ) "
[FAIL] Web Animations: property <max-height> from neutral to [max-content] at (0) should be [calc-size(max-content, 100px + size * 0)]
  assert_equals: expected "calc - size ( max - content , 100px + size * 0 ) " but got "calc - size ( max - content , 100px + 0px + size * 0 ) "
[FAIL] Web Animations: property <max-height> from neutral to [max-content] at (0.3) should be [calc-size(max-content, 70px + size * 0.3)]
  assert_equals: expected "calc - size ( max - content , 70px + size * 0.3 ) " but got "calc - size ( max - content , 70px + 0px + size * 0.3 ) "
[FAIL] Web Animations: property <max-height> from neutral to [max-content] at (0.6) should be [calc-size(max-content, 40px + size * 0.6)]
  assert_equals: expected "calc - size ( max - content , 40px + size * 0.6 ) " but got "calc - size ( max - content , 40px + 0px + size * 0.6 ) "
[FAIL] Web Animations: property <max-height> from neutral to [max-content] at (1) should be [calc-size(max-content, 0px + size * 1)]
  assert_equals: expected "calc - size ( max - content , 0px + size * 1 ) " but got "calc - size ( max - content , 0px + 0px + size * 1 ) "
[FAIL] Web Animations: property <max-height> from neutral to [max-content] at (1.5) should be [calc-size(max-content, -50px + size * 1.5)]
  assert_equals: expected "calc - size ( max - content , - 50px + size * 1.5 ) " but got "calc - size ( max - content , - 50px + 0px + size * 1.5 ) "
Harness: the test ran to completion.
